# Refactor the selection system to use an Image instead of a BitMap