--- a/src/Classes/SelectionMap.gd
+++ b/src/Classes/SelectionMap.gd
@@ -1,6 +1,7 @@
 class_name SelectionMap
 extends Image
 
+var shader: Shader = preload("res://src/Shaders/Invert.shader")
 
 func is_pixel_selected(pixel: Vector2) -> bool:
 	if pixel.x < 0 or pixel.y < 0 or pixel.x >= get_width() or pixel.y >= get_height():
@@ -25,10 +26,14 @@ func clear() -> void:
 
 
 func invert() -> void:
-	for x in get_size().x:
-		for y in get_size().y:
-			var pos := Vector2(x, y)
-			select_pixel(pos, !is_pixel_selected(pos))
+	var params := {
+		"red": true,
+		"green": true,
+		"blue": true,
+		"alpha": true
+	}
+	var gen := ShaderImageEffect.new()
+	gen.generate_image(self, shader, params, get_size())
 
 
 func move_bitmap_values(project, move_offset := true) -> void: