… src/Shaders/ColorSelect.gdshader
… src/Tools/BaseTool.gd
… src/Tools/Bucket.gd
… src/Tools/Draw.gd
… src/Tools/SelectionTools/ColorSelect.gd
… src/Tools/SelectionTools/ColorSelect.tscn
… src/Tools/SelectionTools/EllipseSelect.gd
… src/Tools/SelectionTools/Lasso.gd
… src/Tools/SelectionTools/MagicWand.gd
… src/Tools/SelectionTools/PolygonSelect.gd
… src/Tools/SelectionTools/RectSelect.gd
… src/Tools/SelectionTools/SelectionTool.gd
… src/UI/Canvas/Selection.gd
… src/UI/Dialogs/ImageEffects/DesaturateDialog.gd
… src/UI/Dialogs/ImageEffects/DropShadowDialog.gd
… src/UI/Dialogs/ImageEffects/FlipImageDialog.gd
… src/UI/Dialogs/ImageEffects/GradientDialog.gd
… src/UI/Dialogs/ImageEffects/GradientMapDialog.gd
… src/UI/Dialogs/ImageEffects/HSVDialog.gd
… src/UI/Dialogs/ImageEffects/InvertColorsDialog.gd
… src/UI/Dialogs/ImageEffects/OutlineDialog.gd
… src/UI/Dialogs/ImageEffects/RotateImage.gd