**Requires Godot 3.5 and above. Builds will most likely fail because of this and this PR will not be merged until Godot 3.5-stable is released and Pixelorama's workflows start using it.**

This PR replaces the BitMap that was used to store the selected state of the pixels with an LA8 Image. This massively improves performance, since we get rid of the slow bitmap-to-image conversion. Images were used anyway for the selection marching ants outline to be drawn, and for shaders to use them as masks.

I have also refactor the selection code by making a new `SelectionMap` class that extends Image, and moved some methods from Project.gd to the new class, therefore making the code cleaner and putting the methods to a more correct place.

A comparison of approximate common timings of full canvas selections using the Rectangular Selection tool, between the current and the new method.

 WxH | Rectangular Selection (current) | Rectangular Selection (new) |
| - | - | - |
| 512x512 px | 274 ms | 6 ms |
| 1024x1024 px | 1000 ms | 23 ms |
| 2048x2048 px | 3927 ms | 85 ms |
| 4096x4096 px | 15534 ms | 170 ms |

The Select by Color tool now uses a shader for faster results, and a color similarity option has been added, similar to the same color fill in the Bucket tool. The below approximate timings were tested on blank canvases. Tested on an AMD Ryzen 5 1600 CPU and a NVIDIA GTX 1060 6GB GPU, on openSUSE Tumbleweed.

 WxH | Select by Color (current) | Select by Color (new with shader) |
| - | - | - |
| 512x512 px | 410 ms | 40 ms |
| 1024x1024 px | 1602 ms | 65 ms |
| 2048x2048 px | 6376 ms | 112 ms |
| 4096x4096 px | 26360 ms | 391 ms |